requirements.txt
discord.py>=2.3.0
google-genai>=1.0.0
google-auth>=2.27.0
google-api-python-client>=2.116.0
python-dotenv>=1.0.0
pytest>=8.0.0
aiohttp>=3.9.0
